Username-based login mechanism
The route for login is well-structured
just like the other routes. While the login service that the
controller relies on has too much responsibility and infact its
not clean. It handles if the username doesn't exist,
hashes the password and compares the login pairs if it matches.
It's so unpleasant code and need a better way more cleaner and
abstract way.

Surely I will look in to it later. Clueless.