User identity is bound after verification
Once the user is verified from the verification
middleware, user's identity is bound to the request
object and passed next to the next request handler.
The next request handler which is for sure the
controller will able to easily identify who is the
requester. So no need for either the controller nor
the service to decode the access token and extract
the payload.

One little thing. For those controllers which requires
some parameters to perform action on other user's id has
to implement some sort of authorization to check if the
current requester's privilege is sufficent to perform the
action.