Project status
Currently changed the way of creating a blog post
surely requiring auth and validating the token
through that common validator middleware that assigns
the decoded user token to the request variable
so the blog service layer would know who is the author.

Added the remember me option so no need to create and
persist the refresh token, its just a quick and dirty
implementation of it that invloved in controller and
service.

Now the backend kinda functioning as expected. It's
now on idle state of development. And for now I'm going
back to the frontend and continue the mess I developed in
React.

Any changes needed here would definitely be done and pushed.