Upload profile avatar and role validation
Profile avatar now is stored in file system and
its filename stored in the database, no need to
store the whole blob in the database. The upload
middleware is set properly for multi-handling any
file upload such as blog post thumbnail.

Blog post creation is now validated with the author
role, if the user is an admin or author, otherwise
reject the new blog post.

Prepared a client folder for the frontend build and
redirected any none registered routes to the frontend.

The registered routes now is prefixed with /api, just
to distinguish that these routes are meant for REST api
calls.